pyproject.toml
[project]
name = "s13a"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "comtypes>=1.4.11",
    "dotenv>=0.9.9",
    "google-genai>=1.16.1",
    "mss>=10.0.0",
    "numpy>=2.2.6",
    "onnx>=1.18.0",
    "onnxruntime-gpu>=1.22.0",
    "opencv-python-headless>=4.11.0.86",
    "pillow>=11.2.1",
    "psutil>=7.0.0",
    "pywin32>=310",
    "requests>=2.32.3",
    "rich>=14.0.0",
    "wget>=3.2",
]